--- Model/Config/Source/Order/Status/Pendingpayment.php
@@ -1,20 +1,0 @@
-<?php
-	/**
- * Copyright © 2015 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
-namespace PagoEpayco\Payco\Model\Config\Source\Order\Status;
-
-use Magento\Sales\Model\Order;
-use Magento\Sales\Model\Config\Source\Order\Status;
-
-/**
- * Order Status source model
- */
-class Pendingpayment extends Status
-{
-    /**
-     * @var string[]
-     */
-    protected $_stateStatuses = [Order::STATE_PENDING_PAYMENT];
-}
