+++ PagoEpayco/Payco/CHANGELOG.md
@@ -1,0 +1,17 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+## [1.0.0] - 2021-09-16
+
+### Primera versión
+
+- Se sube la primera versión del plugin de magento
+
+[Unreleased]: https://github.com/epayco/plugin_magento_multistore/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/epayco/plugin_magento_multistore/releases/tag/v1.0.0
