+++ PagoEpayco/Payco/etc/config.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0"?>
+<!--
+/**
+ * Copyright © 2015 Magento. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Store:etc/config.xsd">
+    <default>
+        <payment>
+            <epayco>
+                <active>1</active>
+                <model>PagoEpayco\Payco\Model\Epayco</model>
+                <order_status>pending</order_status>
+                <title>Epayco</title>
+                <p_cust_id_cliente>P_CUST_ID_CLIENTE</p_cust_id_cliente>
+                <p_key>P_KEY</p_key>
+                <public_key>Public_KEY</public_key>
+                <private_key>Private_KEY</private_key>
+                <url_de_respuesta>URL De Respuesta</url_de_respuesta>
+                <url_de_confirmacion>URL de confirmación</url_de_confirmacion>
+                <modo_pruebas>Modo Pruebas</modo_pruebas>
+                <tipo_de_checkout>Tipo de Checkout</tipo_de_checkout>
+                <!--<payment_action>true</payment_action>-->
+                <allowspecific>0</allowspecific>
+                <group>epayco_payment</group>
+            </epayco>
+        </payment>
+    </default>
+</config>
