+++ PagoEpayco/Payco/etc/frontend/routes.xml
