+++ PagoEpayco/Payco/etc/payment.xml
