--- composer.json
@@ -1,22 +1,0 @@
-{
-  "name": "epayco/module-epaycopayment",
-  "description": "Module for payment provide by ePayco",
-  "type": "magento2-module",
-  "license": "GPL-3.0",
-  "authors": [
-    {
-      "name": "ePayco",
-      "email": "ricardo.saldarriaga@payco.co"
-    }
-  ],
-  "minimum-stability": "dev",
-  "require": {},
-  "autoload": {
-    "files": [
-      "registration.php"
-    ],
-    "psr-4": {
-      "PagoEpayco\\Payco\\": ""
-    }
-  }
-}
