--- view/frontend/web/js/view/payment/epayco.js
@@ -1,26 +1,0 @@
-/**
- * Copyright © 2015 Magento. All rights reserved.
- * See COPYING.txt for license details.
- */
-/*browser:true*/
-/*global define*/
-define(
-    [
-        'uiComponent',
-        'Magento_Checkout/js/model/payment/renderer-list'
-    ],
-    function (
-        Component,
-        rendererList
-    ) {
-        'use strict';
-        rendererList.push(
-            {
-                type: 'epayco',
-                component: 'PagoEpayco_Payco/js/view/payment/method-renderer/epayco-method'
-            }
-        );
-        /** Add view logic here if needed */
-        return Component.extend({});
-    }
-);
